Optimalized how TriggerAction works
Performing an action no longer needs using OnSelect before. Thanks to this it was easier to merge two methods into one. Now there in one TriggerAction for both player and AI. Also works in replays.

Added bool in GameManager that indicates if AI is enabled or not. Should be done in UI in future.

It is now possible to watch AI replays.